Fix unit tests, still missing memory alignment issues from Eigen

+++ src/libs/vcl.core/vcl/core/simd/bool16_neon.h
@@ -43,10 +43,10 @@ namespace Vcl
 		VCL_STRONG_INLINE VectorScalar() = default;
 		VCL_STRONG_INLINE VectorScalar(bool s)
 		{
-			mF4[0] = s ? vdupq_n_u32(-1) : vdupq_n_u32(0);
-			mF4[1] = s ? vdupq_n_u32(-1) : vdupq_n_u32(0);
-			mF4[2] = s ? vdupq_n_u32(-1) : vdupq_n_u32(0);
-			mF4[3] = s ? vdupq_n_u32(-1) : vdupq_n_u32(0);
+			mF4[0] = s ? vdupq_n_u32((uint32_t) -1) : vdupq_n_u32(0);
+			mF4[1] = s ? vdupq_n_u32((uint32_t) -1) : vdupq_n_u32(0);
+			mF4[2] = s ? vdupq_n_u32((uint32_t) -1) : vdupq_n_u32(0);
+			mF4[3] = s ? vdupq_n_u32((uint32_t) -1) : vdupq_n_u32(0);
 		}
 		explicit VCL_STRONG_INLINE VectorScalar(uint32x4_t I4_0, uint32x4_t I4_1, uint32x4_t I4_2, const uint32x4_t& I4_3)
 		{
