+++ src/libs/vcl.core/vcl/core/simd/bool4_neon.h
@@ -44,7 +44,7 @@ namespace Vcl
 		VCL_STRONG_INLINE VectorScalar() = default;
 		VCL_STRONG_INLINE VectorScalar(bool s)
 		{
-			_data[0] = s ? vdupq_n_u32(-1) : vdupq_n_u32(0);
+			_data[0] = s ? vdupq_n_u32((uint32_t) -1) : vdupq_n_u32(0);
 		}
 		explicit VCL_STRONG_INLINE VectorScalar(uint32x4_t F4)
 		{
